Chat federation: refactoring code + fix case when video has no chat
